[WIP] trying to get a nice build --watch going
Having a problem with rollup's watch command.

```
yarn build & start-storybook -p 6006
```

This build's with watch, so that storybook can pick up changes in the dist folder.
However, it seems to always be one version of the file behind the current version.
If you save the file, then make a change, then save it again, the dist folder
will still have the build from that initial save... I was not able to figure out why.