Trying to get it working with all kinds of parents
If react is listed as a dev dependency, then it will be installed, and the build contains references to that react.
Then if I link brainstorm into another react app, ega a CRA project, it won't run because
of multiple versions of react.

If it is not installed, then storybook won't run.

To solve this I have moved storybook and its dependencies, including a dev dependency on react,
into a separate folder from the src/ folder (though it still finds stories in the src folder).
This way the main project does not get a version of react, while the storybook does.

But this feels wrong: it feels like I should be able to install react and have webpack
externals take care of it.